backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyPDF2==3.0.1
python-dotenv==1.0.0
openai==1.3.7
Pillow==10.1.0
whitenoise==6.6.0
requests==2.32.4
httpx>=0.24.1 